# public/openapi-spec.yml
openapi: 3.0.0
paths:
...
  /v1/applications:
    post:
      operationId: ApplicationController_createApplication
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectDto"
      responses:
        "200":
          description: ""
      tags:
        - Application
        - Create Application
      security: &a5
        - bearer: []
    get:
      operationId: ApplicationController_createProject
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - Application
        - Get Applications
      security: *a5
info:
  title: Tetrfy
  description: API description
  version: "1.0"
  contact: {}
...
servers: []
components:
  schemas:
...
    UpsertAddressDto:
      type: object
      properties:
        addressId:
          type: string
        name:
          type: string
        address1:
          type: string
        address2:
          type: string
        city:
          type: string
        region:
          type: string
        country:
          type: string
        phone:
          type: string
        phoneType:
          type: object
      required:
        - name
        - address1
        - address2
        - city
        - region
        - country
        - phone
        - phoneType
    UpsertCompanyDto:
      type: object
      properties:
        companyId:
          type: string
        name:
          type: string
        description:
          type: string
        address:
          $ref: "#/components/schemas/UpsertAddressDto"
      required:
        - name
        - description
        - address
    CreateProjectDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        industryId:
          type: string
        company:
          $ref: "#/components/schemas/UpsertCompanyDto"
      required:
        - name
        - description
        - industryId
        - company